+++ src/main/java/persistence/sql/meta/Column.java
@@ -1,10 +1,9 @@
-package persistence.sql.util;
+package persistence.sql.meta;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Transient;
-import persistence.sql.FieldType;
 
 import java.lang.reflect.Field;
 import java.util.Objects;
