+++ src/main/java/persistence/sql/meta/FieldType.java
@@ -1,4 +1,4 @@
-package persistence.sql;
+package persistence.sql.meta;
 
 import java.lang.reflect.Field;
