+++ src/test/java/persistence/sql/dml/DeleteQueryBuilderTest.java
@@ -2,9 +2,9 @@ package persistence.sql.dml;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import persistence.example.Person;
 import persistence.sql.NotEntity;
-import persistence.sql.Person;
-import persistence.sql.Table;
+import persistence.sql.meta.Table;
 
 import static org.assertj.core.api.Assertions.*;
